refactor: ConversationEvents에서 interface로 정의했던 것을 res로 옮김
