+++ src/conversation-events/conversation-events.module.ts
@@ -5,9 +5,11 @@ import { AuthModule } from 'src/auth/auth.module';
 import { RoomsModule } from 'src/rooms/rooms.module';
 import { ConversationEventsService } from './conversation-events.service';
 import { RoomEventModule } from 'src/rooms/item/room-event/room-event.module';
+import { RoomSocketModule } from './room-socket/room-socket.module';
+import { SocketInformationModule } from './socket-information/socket-information.module';
 
 @Module({
-  imports: [UsersModule, AuthModule, RoomsModule, RoomEventModule],
+  imports: [UsersModule, AuthModule, RoomsModule, RoomEventModule, RoomSocketModule, SocketInformationModule],
   providers: [ConversationEventsGateway, ConversationEventsService],
 })
 export class ConversationEventsModule {}
