+++ src/conversation-events/room-socket/room-socket.module.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { RoomSocketService } from './room-socket.service';
+
+@Module({
+  providers: [RoomSocketService],
+})
+export class RoomSocketModule {}
