+++ src/conversation-events/room-socket/room-socket.service.ts
@@ -1,0 +1,4 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class RoomSocketService {}
