+++ src/conversation-events/socket-information/socket-information.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { SocketInformationService } from './socket-information.service';
+import { SocketInformationController } from './socket-information.controller';
+
+@Module({
+  controllers: [SocketInformationController],
+  providers: [SocketInformationService],
+})
+export class SocketInformationModule {}
