+++ src/rooms/item/index.ts
@@ -1,8 +1,8 @@
 import { ChatData } from 'src/conversation-datas/data/chatting';
 import { User } from 'src/users/entities/user.entity';
 import { initRoomEvent, RoomEvent, RoomStatus } from './room-event';
-import { RoomSocket } from 'src/conversation-events/interfaces/room-socket.interface';
-import { SocketInformation } from 'src/conversation-events/interfaces/socket-information.interface';
+import { RoomSocket } from 'src/conversation-events/room-socket';
+import { SocketInformation } from 'src/conversation-events/socket-information';
 
 export class Room implements RoomEvent {
   creatorPk: number;
